move tui function
